+++ src/components/custom/dialogs/TermsDialog.tsx
