+++ src/components/custom/forms/SignupForm.tsx
@@ -2,8 +2,8 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 
-import TermsDialog from '@/components/dialogs/TermsDialog';
-import TermsAgreement from '@/components/feature/TermsAgreement';
+import TermsDialog from '@/components/custom/dialogs/TermsDialog';
+import TermsAgreement from '@/components/custom/feature/TermsAgreement';
 import { Button } from '@/components/ui/button';
 import {
   Form,
