+++ src/components/custom/icons/PlusIcon.tsx
