+++ src/components/ui/button.tsx
@@ -2,7 +2,7 @@ import { Slot } from '@radix-ui/react-slot';
 import { cva, type VariantProps } from 'class-variance-authority';
 import * as React from 'react';
 
-import PlusIcon from '@/components/icons/PlusIcon';
+import PlusIcon from '@/components/custom/icons/PlusIcon';
 import { cn } from '@/lib/utils';
 
 const buttonVariants = cva(
