+++ src/pages/SigninPage.tsx
@@ -1,4 +1,4 @@
-import SigninForm from '@/components/forms/SigninForm';
+import SigninForm from '@/components/custom/forms/SigninForm';
 import SigninLogo from '@/images/signin-logo.svg';
 
 const SigninPage = () => {
