+++ src/pages/VCPage.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useState } from 'react';
 
-import { AudioPlayer } from '@/components/feature/AudioPlayer';
+import { AudioPlayer } from '@/components/custom/feature/AudioPlayer';
 import ProjectTitle from '@/components/section/contents/ProjectTitle';
 import VCMainContents from '@/components/section/contents/VCMainContents';
 import VCOptionsSidebar from '@/components/section/sidebar/VCSidebar';
